Add fragment warning code from Apollo client.
Also normalize document strings (for whitespace anyway).

Note that we are simply checking fragment strings for equality, not ensuring that fragment definitions within the larger graphql document have exact equality.

We could do the above, but it'd be quite a bit more complicated, and I'm not sure if there's much benefit. AFAICT the main reason for wanting that before was exactly this functionality.