Add support for circular dependencies using thunks
When colocating fragments, circular dependencies often become an issue.
While GraphQL doesn't support circular fragment dependencies, it is
still possible to have circular dependencies in the component files that
declare the fragments.

An example of this is when rendering a tree structure using GraphQL.
While GraphQL doesn't allow for recursive queries, it is still possible
to query an array of descendants and build a tree data structure in the
client. This data structure can then be rendered recursively with React.